chore: move code to the correct file
