+++ frappe/boot.py
@@ -528,7 +528,7 @@ def get_sentry_dsn():
 
 
 def get_sidebar_items():
-	from frappe.utils.install import auto_generate_sidebar_from_module
+	from frappe.desk.doctype.workspace_sidebar.workspace_sidebar import auto_generate_sidebar_from_module
 
 	sidebars = frappe.get_all(
 		"Workspace Sidebar", fields=["name", "header_icon"], filters={"name": ["not like", "%My Workspaces%"]}
